requirements.txt
python-dotenv
fastapi
uvicorn
ibapi
pydantic
sqlalchemy
mysql-connector-python
pymysql